+++ lib/ts/recipe/thirdparty/types.ts
@@ -100,10 +100,7 @@ export type RecipeInterface = {
     signInUp(input: {
         thirdPartyId: string;
         thirdPartyUserId: string;
-        email: {
-            id: string;
-            isVerified: boolean;
-        };
+        email: string;
         userContext: any;
     }): Promise<{ status: "OK"; createdNewUser: boolean; user: User }>;
 };