+++ lib/ts/recipe/thirdpartyemailpassword/recipeImplementation/thirdPartyRecipeImplementation.ts
@@ -23,10 +23,7 @@ export default function getRecipeInterface(recipeInterface: ThirdPartyEmailPassw
         signInUp: async function (input: {
             thirdPartyId: string;
             thirdPartyUserId: string;
-            email: {
-                id: string;
-                isVerified: boolean;
-            };
+            email: string;
             userContext: any;
         }): Promise<{ status: "OK"; createdNewUser: boolean; user: User }> {
             let result = await recipeInterface.thirdPartySignInUp(input);