+++ lib/ts/recipe/thirdpartyemailpassword/types.ts
@@ -107,10 +107,7 @@ export type RecipeInterface = {
     thirdPartySignInUp(input: {
         thirdPartyId: string;
         thirdPartyUserId: string;
-        email: {
-            id: string;
-            isVerified: boolean;
-        };
+        email: string;
         userContext: any;
     }): Promise<{ status: "OK"; createdNewUser: boolean; user: User }>;